Ensure that WaitForAPIServerReady always re-dials through the loadbalancer
(cherry picked from commit e763fadbba5fe933c5111a41c7af0f58fdd9814a)
